publish two components: HighTable and ControlledHighTable
